Moved persistence classes
